telegram_bot/requirements.txt
aiogram>=3.3
psycopg2-binary>=2.9
python-dotenv>=1.0
aiohttp>=3.9
aiofiles>=23.0
